streamlit/requirements.txt
streamlit
requests
google-cloud-firestore